env variables and index.tsx for api call
