+++ frontend/.eslintrc.json
@@ -1,3 +1,3 @@
 {
-  "extends": ["next/core-web-vitals"]
+  "extends": ["next/babel", "next/core-web-vitals"]
 }
